pyproject.toml
[project]
name = "gpt-pr-review-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "httpx>=0.28.1",
    "openai>=1.102.0",
    "pydantic-settings>=2.10.1",
    "pyyaml>=6.0.2",
]

[dependency-groups]
dev = [
    "pre-commit>=4.3.0",
    "pytest>=8.4.1",
    "ruff>=0.12.11",
]

[tool.uv.scripts]
dev = "uvicorn app.main:app --reload --port 8000"
test = "pytest -v"
lint = "ruff check ."
fmt = "ruff format ."
review = "python -m app.cli_review"
summary = "python -m tools.ci_summary"
status = "python -m tools.ci_status"
